# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "招聘中心系统 API 文档",
        "title": "招聘中心 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/companies/{companyId}/jobs": {
            "get": {
                "description": "获取指定公司发布的所有职位",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "获取公司职位",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公司ID",
                        "name": "companyId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.JobListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/jobs": {
            "get": {
                "description": "分页获取职位列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "获取职位列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.JobListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建新职位,需要提供职位相关信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "创建职位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "职位创建请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.JobResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/expired": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取所有已过期的职位列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "获取已过期职位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.JobListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/search": {
            "get": {
                "description": "根据关键词搜索职位信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "关键词搜索职位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.JobListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/search/condition": {
            "post": {
                "description": "根据多个条件筛选职位",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "多条件搜索职位",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "description": "搜索条件",
                        "name": "conditions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.JobListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/{id}": {
            "get": {
                "description": "获取指定ID的职位详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "获取职位详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "职位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.JobResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新指定ID的职位信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "更新职位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "职位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "职位更新请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Job"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "职位不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定ID的职位",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "删除职位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "职位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "职位不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "enums.JobType": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-comments": {
                "JobTypeFullTime": "全职",
                "JobTypeIntern": "实习",
                "JobTypePartTime": "兼职"
            },
            "x-enum-descriptions": [
                "全职",
                "兼职",
                "实习"
            ],
            "x-enum-varnames": [
                "JobTypeFullTime",
                "JobTypePartTime",
                "JobTypeIntern"
            ]
        },
...
        "model.Job": {
            "type": "object",
            "properties": {
                "applyCount": {
                    "description": "申请次数",
                    "type": "integer"
                },
                "companyId": {
                    "type": "integer"
                },
                "createTime": {
                    "type": "string"
                },
                "deleteStatus": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "jobBenefit": {
                    "type": "string"
                },
                "jobCategory": {
                    "type": "string"
                },
                "jobContact": {
                    "type": "string"
                },
                "jobDescribe": {
                    "type": "string"
                },
                "jobEducation": {
                    "type": "string"
                },
                "jobExperience": {
                    "type": "string"
                },
                "jobExpireTime": {
                    "type": "string"
                },
                "jobLocation": {
                    "type": "string"
                },
                "jobSalary": {
                    "type": "string"
                },
                "jobSkill": {
                    "type": "string"
                },
                "jobSource": {
                    "type": "string"
                },
                "jobType": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "description": "优先级",
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "description": "职位标签",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updateTime": {
                    "type": "string"
                },
                "viewCount": {
                    "description": "添加新的字段和方法",
                    "type": "integer"
                }
            }
        },
...
        "request.CreateJobRequest": {
            "type": "object"
        },
...
        "request.UpdateJobRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "jobBenefit": {
                    "type": "string",
                    "maxLength": 500
                },
                "jobCategory": {
                    "type": "string",
                    "maxLength": 50
                },
                "jobContact": {
                    "type": "string",
                    "maxLength": 100
                },
                "jobDescribe": {
                    "type": "string",
                    "maxLength": 2000
                },
                "jobEducation": {
                    "type": "string",
                    "maxLength": 50
                },
                "jobExperience": {
                    "type": "string",
                    "maxLength": 50
                },
                "jobExpireTime": {
                    "type": "string"
                },
                "jobLocation": {
                    "type": "string",
                    "maxLength": 200
                },
                "jobSalary": {
                    "type": "string",
                    "maxLength": 50
                },
                "jobSkill": {
                    "type": "string",
                    "maxLength": 500
                },
                "jobSource": {
                    "type": "string",
                    "maxLength": 100
                },
                "jobType": {
                    "$ref": "#/definitions/enums.JobType"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
        "response.JobListResponse": {
            "type": "object",
            "properties": {
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.JobResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.JobResponse": {
            "type": "object",
            "properties": {
                "applyCount": {
                    "type": "integer"
                },
                "companyId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "jobBenefit": {
                    "type": "string"
                },
                "jobCategory": {
                    "type": "string"
                },
                "jobContact": {
                    "type": "string"
                },
                "jobDescribe": {
                    "type": "string"
                },
                "jobEducation": {
                    "type": "string"
                },
                "jobExperience": {
                    "type": "string"
                },
                "jobExpireTime": {
                    "type": "string"
                },
                "jobLocation": {
                    "type": "string"
                },
                "jobSalary": {
                    "type": "string"
                },
                "jobSkill": {
                    "type": "string"
                },
                "jobSource": {
                    "type": "string"
                },
                "jobType": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "viewCount": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/errors.ErrorCode"
                        }
                    ]
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应信息",
                    "type": "string"
                }
            }
        }
    }
}